refactor(progress-bar): streamline type definitions and enhance prop interfaces
- Consolidated type definitions for the ProgressBar component by moving relevant props to a new slots file.
- Updated the ProgressBarProps interface to exclude unnecessary Chakra UI props, improving clarity and usability.
- Enhanced type safety by integrating Aria props and custom data attributes into the root props interface.
